requirements.txt
python-dotenv==0.21.0
python-telegram-bot==13.0
google-cloud-dialogflow==2.17.0
google-cloud-storage==2.7.0
vk_api==11.9.9

